Refactor: modify session handling component
